add login and logout functionality

+++ base/views.py
@@ -1,9 +1,38 @@
+from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
 from django.shortcuts import render, redirect
 from django.db.models import Q
 from .models import Room, Topic
 from .forms import RoomForm
 
+def login_page(request):
+    if request.user.is_authenticated:
+        return redirect('home')
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exists.')
+            
+        user = authenticate(request, username=username, password=password)
 
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'password not correct.')
+            
+    context = {}
+    return render(request, 'base/login_register.html', context)
+
+def logout_user(request):
+    logout(request)
+    return redirect('home')
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
     # icontains represent type of filter (i) for case insensetive and contain for match with part of the string
@@ -24,7 +53,7 @@ def room(request, pk):
 
     return render(request, "base/room.html", context)
 
-
+@login_required(login_url='login')
 def create_room(request):
     form = RoomForm()
     if request.method == "POST":
@@ -38,9 +67,12 @@ def create_room(request):
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url='login')
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here")
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -50,8 +82,11 @@ def update_room(request, pk):
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url='login')
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here")
     if request.method == "POST":
         room.delete()
         return redirect("home")
